requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.3
djangorestframework==3.14.0
pytz==2023.3
secure-smtplib==0.1.1
sqlparse==0.4.4
typing_extensions==4.7.1
